clean repo, minor refactor
