+++ src/app/app.component.html
@@ -3,13 +3,4 @@
   <router-outlet></router-outlet>
 </div>
 
-<footer class="d-flex flex-wrap justify-content-between align-items-center py-3 my-4 border-top p-3">
-  <div class="col-md-4 d-flex align-items-center">
-    <span class="mb-3 mb-md-0 text-muted">© 2023 <a target="_blank" href="https://github.com/Quethzel"
-        class="text-muted">Quethzel Diaz</a></span>
-  </div>
-  <a href="/home" class="mb-3 me-2 mb-md-0 text-muted text-decoration-none lh-1">My Money Coach</a>
-  <ul class="nav col-md-4 justify-content-end list-unstyled d-flex">
-    <li class="ms-3 text-muted">{{version}}</li>
-  </ul>
-</footer>
+<app-footer [version]="version"></app-footer>
