--- src/app/app.component.scss
@@ -1,10 +1,0 @@
-footer {
-    font-size: 13px;
-    position: absolute;
-    width: 100%;
-    margin: 0 !important;
-    bottom: 0;
-}
-footer a {
-    text-decoration: none;
-}
