--- src/app/components/city-cell-editor/city-cell-editor.component.html
@@ -1,5 +1,0 @@
-<select class="form-select" #input [(ngModel)]="value">
-    <option *ngFor="let city of cities" 
-        value="{{city.code}}">{{city.name}}
-    </option>
-</select>
