--- src/app/components/city-cell-renderer/city-cell-renderer.component.html
@@ -1,1 +1,0 @@
-{{value}}
