+++ src/app/components/expenses/expenses.component.html
@@ -34,5 +34,5 @@
     <div style="height: 500px; overflow-y: auto;">
         <app-editable-expense-grid [expenses]="expenses" (saveItem)="saveItem($event)" (deleteItem)="deleteItem($event)"></app-editable-expense-grid>
     </div>
-
+    
 </div>
