+++ src/app/components/user-profile/user-profile.component.html
@@ -1,17 +1,8 @@
 <!-- <div class="d-flex flex-column"> -->
 <div class="container-fluid">
     <h3 class="text-dark mb-4">Profile</h3>
-    <div class="row mb-3">
-        <div class="col-lg-4">
-            <div class="card mb-3">
-                <div class="card-body text-center shadow"><img class="rounded-circle mb-3 mt-4"
-                        src="assets/img/image1.jpg" width="160" height="160">
-                    <div class="mb-3"><button class="btn btn-primary" type="button">Change Photo</button>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="col-lg-8">
+    <div class="row mb-3 d-flex justify-content-center align-items-center">
+        <div class="col-sm-10">
             <div class="accordion" id="accordionPanelsStayOpenExample">
                 <div class="accordion-item">
                     <h2 class="accordion-header">
