--- src/app/components/variable-expenses-kpi-bar/variable-expenses-kpi-bar.component.html
@@ -1,26 +1,0 @@
-<div class="kpi-bar">
-    <!-- <div class="kpi-box">
-        <p class="title">Total Expenses</p>
-        <h1>$148,197.21</h1>
-    </div> -->
-    <div class="kpi-box">
-        <p class="title">Monthly Expenses</p>
-        <h1>{{totalMonthlyExpenses | currency}}</h1>
-    </div>
-    <div class="kpi-box">
-        <p class="title">Average Daily Expenses</p>
-        <h1>$4,897.99</h1>
-    </div>
-    <div class="kpi-box">
-        <p class="title">Remaning Monthly Budget</p>
-        <h1>$21,850.00</h1>
-    </div>
-    <div class="kpi-box">
-        <p class="title">Days left</p>
-        <h1>13</h1>
-    </div>
-    <div class="kpi-box">
-        <p class="title">Total Rows</p>
-        <h1>{{expenses.length}}</h1>
-    </div>
-</div>
