--- src/app/components/variable-expenses-kpi-bar/variable-expenses-kpi-bar.component.scss
@@ -1,19 +1,0 @@
-.kpi-bar {
-    display: flex;
-    flex-wrap: wrap;
-}
-.kpi-box {
-    // flex: auto;
-    flex: 10%;
-    padding: 5px;
-    text-align: center;
-    margin: 5px;
-    background-color: #E7E7E7;
-}
-.kpi-box h1 {
-    font-size: 22px;
-    font-weight: 400px;
-}
-.title {
-    font-size: 12px;
-}
