+++ src/app/forgot-pass/forgot-pass.component.html
@@ -5,7 +5,7 @@
                 <div class="card-body p-0">
                     <div class="row">
                         <div class="col-lg-6 d-none d-lg-flex">
-                            <img class="img-fluid" src="../../assets/img/image3.jpg">
+                            <img class="img-fluid" src="/assets/img/forgot-pass.jpg">
                         </div>
                         <div class="col-lg-6">
                             <div class="p-5">
