+++ src/app/home/footer/footer.component.html
@@ -1,1 +1,10 @@
-<p>footer works!</p>
+<footer class="d-flex flex-wrap justify-content-between align-items-center py-3 my-4 border-top p-3">
+    <div class="col-md-4 d-flex align-items-center">
+      <span class="mb-3 mb-md-0 text-muted">© 2023 <a target="_blank" href="https://github.com/Quethzel"
+          class="text-muted">Quethzel Diaz</a></span>
+    </div>
+    <a href="/home" class="mb-3 me-2 mb-md-0 text-muted text-decoration-none lh-1">My Money Coach</a>
+    <ul class="nav col-md-4 justify-content-end list-unstyled d-flex">
+      <li class="ms-3 text-muted">{{version}}</li>
+    </ul>
+  </footer>
