+++ src/app/home/footer/footer.component.scss
@@ -1,0 +1,10 @@
+footer {
+    font-size: 13px;
+    position: absolute;
+    width: 100%;
+    margin: 0 !important;
+    bottom: 0;
+}
+footer a {
+    text-decoration: none;
+}
