+++ src/app/home/footer/footer.component.ts
@@ -1,10 +1,11 @@
-import { Component } from '@angular/core';
-//TODO: This component is not used
+import { Component, Input } from '@angular/core';
+
 @Component({
   selector: 'app-footer',
   templateUrl: './footer.component.html',
   styleUrls: ['./footer.component.scss']
 })
 export class FooterComponent {
+  @Input() version: string;
 
 }
