+++ src/app/register/register.component.html
@@ -3,7 +3,7 @@
         <div class="card-body p-0">
             <div class="row">
                 <div class="col-lg-5 d-none d-lg-flex">
-                    <img class="img-fluid" src="../../assets/img/image1.jpg">
+                    <img class="img-fluid" src="/assets/img/login.jpg">
                 </div>
                 <div class="col-lg-7">
                     <div class="p-5">
