+++ src/app/services/user.service.ts
@@ -2,14 +2,21 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { environment } from 'src/environments/environment';
 import { IUser } from '../models/interfaces/IUser';
+import { IUserCity } from '../models/interfaces/ICity';
+import { of } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
 })
 export class UserService {
   private apiURL = `${environment.mcApi}/user`;
+  private cities: IUserCity[];
 
   constructor(private http: HttpClient) { }
+
+  get cties() {
+    return this.cities;
+  }
 
   getInfo() {
     return this.http.get<IUser>(this.apiURL);
@@ -19,8 +26,30 @@ export class UserService {
     return this.http.put(this.apiURL, user);
   }
 
+  //TODO: delete account is not implemented
   deleteAccount() {
 
+  }
+
+  loadCities() {
+    this.cities = [
+      { code: 'MTY', name: 'Monterrey', default: true },
+      { code: 'CDMX', name: 'CDMX', default: false },
+      { code: 'XAL', name: 'Xalapa', default: false },
+      { code: 'OAX', name: 'Oaxaca', default: false },
+    ];
+
+    return of(this.cities);
+  }
+
+  getDefaultCity() {
+    let city = this.cities.find(c => c.default == true);
+    if (!city) city = this.cities[0];
+    return city;
+  }  
+
+  getCityCodes() {
+    return this.cities.map(c => c.code);
   }
 
 }
